+++ swing/framework-ui/src/main/java/is/codion/swing/framework/ui/EntityDialogs.java
@@ -329,7 +329,7 @@ public final class EntityDialogs {
 				throw new IllegalArgumentException("All entities must be of the same type when editing");
 			}
 
-			ComponentValue<T, ?> componentValue = editComponentFactory.componentValue(editModel, defaultValue.apply(entities));
+			ComponentValue<T, ?> componentValue = editComponentFactory.component(editModel, defaultValue.apply(entities));
 			inputDialog(componentValue)
 							.owner(owner)
 							.location(location)