+++ swing/framework-ui/src/main/java/is/codion/swing/framework/ui/component/DefaultEditComponentFactory.java
@@ -66,7 +66,7 @@ public class DefaultEditComponentFactory<T, C extends JComponent> implements Edi
 	}
 
 	@Override
-	public ComponentValue<T, C> componentValue(SwingEntityEditModel editModel, T value) {
+	public ComponentValue<T, C> component(SwingEntityEditModel editModel, T value) {
 		requireNonNull(editModel);
 		if (attribute instanceof ForeignKey) {
 			return createForeignKeyComponentValue((ForeignKey) attribute, editModel, (Entity) value);