+++ changelog.md
@@ -64,6 +64,7 @@ Codion Change Log
 - EntityApplicationPanel no longer abstract.
 - EntityEditPanel constructors now protected.
 - EntityEditPanel.insertCommandBuilder(), updateCommandBuilder() and deleteCommandBuilder() renamed insertCommand(), updateCommand() and deleteCommand().
+- EntityComponentFactory.componentValue() renamed component().
 ### is.codion.plugin.flatlaf.intellij
 - IntelliJThemeProvider, disable FlatLaf logging due to FlatLaf #990.
 - material.Dracula theme renamed DraculaMaterial, conflicted with the other Dracula theme.